Refactor bonus overview into tabbed statistics view
- Remove standalone 'b' keyboard shortcut for bonus overview
- Integrate bonus overview as second tab within statistics view ('s' key)
- Add centered layout for bonus progress display
- Implement tab navigation with Tab/Arrow keys or number keys (1/2)
- Add keyboard shortcuts help text in notification bar
- Maintain all bonus calculation functionality within unified interface

Navigation Flow:
- Press 's' to open Statistics view
- Tab/←→ or 1/2 to switch between Monthly Stats and Bonus Overview
- Bonus tab only appears when bonus tracking is enabled
- Both views centered and properly formatted
