+++ source/components/StatisticsView.tsx
@@ -7,6 +7,11 @@ import {
 	StatisticsUseCase,
 	type YearlyStatistics,
 } from '../use-cases/StatisticsUseCase.js';
+import {
+	BonusCalculator,
+	type BonusProgress,
+	type TierVisualization,
+} from '../bonus/BonusCalculator.js';
 import {NotificationBar} from './NotificationBar.js';
 import {StatisticsGrid} from './StatisticsGrid.js';
 
@@ -15,10 +20,19 @@ export type StatisticsViewProps = {
 	config: JiraConfig;
 };
 
+type StatisticsTab = 'monthly' | 'bonus';
+
 export function StatisticsView({onBack, config}: StatisticsViewProps) {
 	const [statistics, setStatistics] = useState<YearlyStatistics | undefined>(
 		undefined,
 	);
+	const [bonusProgress, setBonusProgress] = useState<BonusProgress | undefined>(
+		undefined,
+	);
+	const [tierVisualizations, setTierVisualizations] = useState<
+		TierVisualization[]
+	>([]);
+	const [activeTab, setActiveTab] = useState<StatisticsTab>('monthly');
 	const [isLoading, setIsLoading] = useState(true);
 	const [error, setError] = useState<string | undefined>(undefined);
 
@@ -41,6 +55,14 @@ export function StatisticsView({onBack, config}: StatisticsViewProps) {
 		return new StatisticsUseCase(jiraClient, attendanceManager, config.bonus);
 	}, [jiraClient, attendanceManager, config.bonus]);
 
+	const bonusCalculator = useMemo(() => {
+		if (!config.bonus?.enabled) {
+			return undefined;
+		}
+
+		return new BonusCalculator(config.bonus);
+	}, [config.bonus]);
+
 	// Load statistics data
 	useEffect(() => {
 		async function loadStatistics() {
@@ -55,6 +77,19 @@ export function StatisticsView({onBack, config}: StatisticsViewProps) {
 				setError(undefined);
 				const stats = await statisticsUseCase.execute();
 				setStatistics(stats);
+
+				// Load bonus data if calculator is available
+				if (bonusCalculator && stats.totalHours) {
+					const progress = bonusCalculator.calculateBonusProgress(
+						stats.totalHours,
+					);
+					setBonusProgress(progress);
+
+					const visualizations = bonusCalculator.getTierVisualizations(
+						progress.currentBonusDays,
+					);
+					setTierVisualizations(visualizations);
+				}
 			} catch (error_: unknown) {
 				const message =
 					error_ instanceof Error ? error_.message : String(error_);
@@ -65,12 +100,21 @@ export function StatisticsView({onBack, config}: StatisticsViewProps) {
 		}
 
 		void loadStatistics();
-	}, [statisticsUseCase]);
+	}, [statisticsUseCase, bonusCalculator]);
 
 	// Handle keyboard input
 	useInput((input, key) => {
 		if (key.escape || input === 'q') {
 			onBack();
+		} else if (key.tab || key.leftArrow || key.rightArrow) {
+			// Switch between tabs
+			setActiveTab(currentTab =>
+				currentTab === 'monthly' ? 'bonus' : 'monthly',
+			);
+		} else if (input === '1') {
+			setActiveTab('monthly');
+		} else if (input === '2' && bonusCalculator) {
+			setActiveTab('bonus');
 		}
 	});
 
@@ -103,13 +147,211 @@ export function StatisticsView({onBack, config}: StatisticsViewProps) {
 		);
 	}
 
+	const showBonusTab = Boolean(bonusCalculator && bonusProgress);
+
 	return (
 		<Box flexDirection="column">
-			<StatisticsGrid statistics={statistics} bonusConfig={config.bonus} />
+			{/* Tab Navigation */}
+			<Box marginBottom={1} justifyContent="center">
+				<Box marginRight={2}>
+					<Text
+						color={activeTab === 'monthly' ? 'cyan' : 'gray'}
+						bold={activeTab === 'monthly'}
+					>
+						{activeTab === 'monthly'
+							? '[ 1. Monthly Stats ]'
+							: '  1. Monthly Stats  '}
+					</Text>
+				</Box>
+				{showBonusTab && (
+					<Box>
+						<Text
+							color={activeTab === 'bonus' ? 'cyan' : 'gray'}
+							bold={activeTab === 'bonus'}
+						>
+							{activeTab === 'bonus'
+								? '[ 2. Bonus Overview ]'
+								: '  2. Bonus Overview  '}
+						</Text>
+					</Box>
+				)}
+			</Box>
+
+			{/* Tab Content */}
+			{activeTab === 'monthly' ? (
+				<StatisticsGrid statistics={statistics} bonusConfig={config.bonus} />
+			) : (
+				showBonusTab && renderBonusOverview()
+			)}
 
 			<Box marginTop={1}>
-				<NotificationBar notifications={[]} />
+				<NotificationBar
+					notifications={[
+						{
+							id: 1,
+							message: showBonusTab
+								? '[Tab/←→] Switch Tabs [1] Monthly [2] Bonus [Q] Back'
+								: '[Q] Back',
+							type: 'info',
+						},
+					]}
+				/>
 			</Box>
 		</Box>
 	);
+
+	function renderBonusOverview() {
+		if (!bonusProgress || !config.bonus) {
+			return null;
+		}
+
+		const targetYear = new Date().getFullYear();
+
+		return (
+			<Box flexDirection="column" alignItems="center">
+				{/* Current Status */}
+				<Box flexDirection="column" marginBottom={1} alignItems="center">
+					<Text bold color="cyan">
+						Bonus Progress {targetYear}
+					</Text>
+					<Text>═══════════════════════════════════════════════════</Text>
+				</Box>
+
+				<Box flexDirection="column" marginBottom={1} alignItems="center">
+					<Text>
+						Current Status: <Text bold>{bonusProgress.currentBonusDays}</Text>{' '}
+						Bonus Days
+					</Text>
+					<Text>
+						├─ Target Progress:{' '}
+						<Text bold>
+							{bonusProgress.currentBonusDays} / {config.bonus.targetDays} (
+							{Math.round(
+								(bonusProgress.currentBonusDays / config.bonus.targetDays) *
+									100,
+							)}
+							%)
+						</Text>
+					</Text>
+					<Text>
+						├─ Current Tier:{' '}
+						<Text bold color="yellow">
+							{bonusProgress.currentTier.name}
+						</Text>{' '}
+						({(bonusProgress.currentTier.rate * 100).toFixed(1)}% per day)
+					</Text>
+					<Text>
+						├─ Earned Bonus:{' '}
+						<Text bold color="green">
+							{bonusProgress.earnedBonusPercentage.toFixed(2)}%
+						</Text>
+					</Text>
+					<Text>
+						└─ Projected Year-End:{' '}
+						<Text bold>
+							{bonusProgress.projectedYearEnd} days (
+							{Math.round(
+								(bonusProgress.projectedYearEnd / config.bonus.targetDays) *
+									100,
+							)}
+							%)
+						</Text>
+					</Text>
+				</Box>
+
+				{/* Tier Progress */}
+				<Box flexDirection="column" marginBottom={1} alignItems="center">
+					<Text bold>Tier Progress:</Text>
+					<Box flexDirection="column" marginLeft={2}>
+						{tierVisualizations.map(viz => (
+							<Box key={viz.tier.name} marginBottom={0}>
+								<Text>
+									{viz.tier.name}{' '}
+									{renderProgressBar(viz.percentage, viz.isCurrent)}{' '}
+									{viz.isCompleted ? (
+										<Text color="green">✓ Complete</Text>
+									) : (
+										<Text>
+											{viz.progress.toFixed(1)}/
+											{viz.total === Number.POSITIVE_INFINITY
+												? `${viz.tier.startDay}+`
+												: viz.total}{' '}
+											({viz.percentage.toFixed(0)}%)
+										</Text>
+									)}
+								</Text>
+							</Box>
+						))}
+					</Box>
+				</Box>
+
+				{/* Next Milestone */}
+				{bonusProgress.nextMilestone && (
+					<Box flexDirection="column" marginBottom={1} alignItems="center">
+						<Text bold>Next Milestone:</Text>
+						<Box marginLeft={2}>
+							<Text>
+								• {bonusProgress.nextMilestone.targetDays} days -{' '}
+								{bonusProgress.nextMilestone.name} (
+								<Text color="yellow">
+									{bonusProgress.nextMilestone.daysRemaining} days to go
+								</Text>
+								)
+							</Text>
+						</Box>
+					</Box>
+				)}
+
+				{/* Key Milestones Summary */}
+				<Box flexDirection="column" marginBottom={1} alignItems="center">
+					<Text bold>Key Milestones:</Text>
+					<Box marginLeft={2}>
+						<Text>
+							{bonusProgress.currentBonusDays >= 120 ? '✓' : '•'} 120 days -
+							Tier 2 starts
+							{bonusProgress.currentBonusDays < 120 &&
+								` (${
+									Math.round((120 - bonusProgress.currentBonusDays) * 10) / 10
+								} days to go)`}
+						</Text>
+						<Text>
+							{bonusProgress.currentBonusDays >= 160 ? '✓' : '•'} 160 days -
+							Tier 3 starts
+							{bonusProgress.currentBonusDays < 160 &&
+								` (${
+									Math.round((160 - bonusProgress.currentBonusDays) * 10) / 10
+								} days to go)`}
+						</Text>
+						<Text>
+							{bonusProgress.currentBonusDays >= config.bonus.targetDays
+								? '✓'
+								: '•'}{' '}
+							{config.bonus.targetDays} days - 100% Target
+							{bonusProgress.currentBonusDays < config.bonus.targetDays &&
+								` (${
+									Math.round(
+										(config.bonus.targetDays - bonusProgress.currentBonusDays) *
+											10,
+									) / 10
+								} days to go)`}
+						</Text>
+					</Box>
+				</Box>
+			</Box>
+		);
+	}
+
+	function renderProgressBar(
+		percentage: number,
+		isCurrent: boolean,
+		width = 20,
+	): string {
+		const filled = Math.round((percentage / 100) * width);
+		const empty = width - filled;
+
+		const filledChar = isCurrent ? '█' : '▓';
+		const emptyChar = '░';
+
+		return `[${filledChar.repeat(filled)}${emptyChar.repeat(empty)}]`;
+	}
 }
